RE-38: split project struct on microservices
