+++ cmd/main.go
@@ -8,6 +8,8 @@ import (
 
 	"retarget/configs"
 	authApp "retarget/internal/auth-service/app"
+
+	// profileApp "retarget/internal/profile-service/app"
 	mailApp "retarget/internal/mail-service/app"
 )
 
@@ -19,6 +21,8 @@ func main() {
 
 	go authApp.Run(cfg)
 	go mailApp.Run(cfg)
+	// go advApp.Run(cfg)
+	// go profileApp.Run(cfg)
 
 	stop := make(chan os.Signal, 1)
 	signal.Notify(stop, os.Interrupt, syscall.SIGTERM)
