--- internal/controller/http/auth/logoutHandler.go
@@ -1,45 +1,0 @@
-package auth
-
-import (
-	"encoding/json"
-	"net/http"
-	"retarget/internal/entity"
-	"time"
-)
-
-func (c *AuthController) LogoutHandler(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		w.WriteHeader(http.StatusMethodNotAllowed)
-		json.NewEncoder(w).Encode(entity.NewResponse(true, "Method Not Allowed"))
-		return
-	}
-
-	cookie, err := r.Cookie("session_id")
-	if err != nil {
-		w.WriteHeader(http.StatusUnprocessableEntity)
-		json.NewEncoder(w).Encode(entity.NewResponse(true, err.Error()))
-		return
-	}
-
-	expires := time.Unix(0, 0)
-	http.SetCookie(w, &http.Cookie{
-		Name:     "session_id",
-		Value:    "",
-		Expires:  expires,
-		HttpOnly: true,
-		Secure:   false,
-		SameSite: http.SameSiteLaxMode,
-		Path:     "/",
-		MaxAge:   -1,
-	})
-
-	err = c.authUsecase.Logout(cookie.Value)
-	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		json.NewEncoder(w).Encode(entity.NewResponse(true, err.Error()))
-		return
-	}
-
-	w.WriteHeader(http.StatusOK)
-	w.Write([]byte("Logout successful"))
-}
