+++ internal/controller/http/router.go
@@ -1,12 +1,10 @@
 package http
 
 import (
-	handlerAuth "retarget/internal/controller/http/auth"
-
+	usecaseAuth "retarget/internal/auth-service/usecase/auth"
 	handlerBanner "retarget/internal/controller/http/banner"
-	logs "retarget/internal/controller/http/middleware"
-	usecaseAuth "retarget/internal/usecase/auth"
 	usecaseBanner "retarget/internal/usecase/banner"
+	logs "retarget/pkg/middleware"
 
 	"github.com/gorilla/mux"
 )
@@ -14,10 +12,7 @@ import (
 func SetupRoutes(authUsecase *usecaseAuth.AuthUsecase, bannerUsecase *usecaseBanner.BannerUsecase) *mux.Router {
 	r := mux.NewRouter()
 
-	r.Use(logs.ErrorMiddleware)
-
-	authRoutes := handlerAuth.SetupAuthRoutes(authUsecase)
-	r.PathPrefix("/api/v1/auth/").Handler(authRoutes)
+	r.Use(logs.LogMiddleware)
 
 	bannerRoutes := handlerBanner.SetupBannerRoutes(authUsecase, bannerUsecase)
 	r.PathPrefix("/api/v1/banner/").Handler(bannerRoutes)
