--- internal/entity/session.go
@@ -1,20 +1,0 @@
-package entity
-
-import "time"
-
-type Session struct {
-	ID        string    `json:"id"`
-	UserID    int       `json:"user_id"`
-	Expires   time.Time `json:"expires"`
-	CreatedAt time.Time `json:"created_at"`
-}
-
-var ErrSessionNotFound = &Error{"Session not found"}
-
-type Error struct {
-	Msg string
-}
-
-func (e *Error) Error() string {
-	return e.Msg
-}
