+++ internal/mail-service/controller/http/router.go
@@ -1,9 +1,9 @@
 package http
 
 import (
-	md "retarget/internal/controller/http/middleware"
 	handlerMail "retarget/internal/mail-service/controller/http/mail"
 	usecaseMail "retarget/internal/mail-service/usecase/mail"
+	middleware "retarget/pkg/middleware"
 
 	"github.com/gorilla/mux"
 )
@@ -11,7 +11,7 @@ import (
 func SetupRoutes(mailUsecase *usecaseMail.MailUsecase) *mux.Router {
 	r := mux.NewRouter()
 
-	r.Use(md.ErrorMiddleware)
+	r.Use(middleware.LogMiddleware)
 
 	mailRoutes := handlerMail.SetupMailRoutes(mailUsecase)
 	r.PathPrefix("/api/v1/mail/").Handler(mailRoutes)
